Rename kubernetes module
