+++ instrumentation/kubernetes-client-7.0/kubernetes-client-7.0.gradle
