+++ settings.gradle
@@ -106,7 +106,7 @@ include ':instrumentation:jsp-2.3'
 include ':instrumentation:kafka-clients-0.11'
 include ':instrumentation:kafka-streams-0.11'
 include ':instrumentation:khttp-0.1'
-include ':instrumentation:kubernetes'
+include ':instrumentation:kubernetes-client-7.0'
 include ':instrumentation:lettuce:lettuce-4.0'
 include ':instrumentation:lettuce:lettuce-5.0'
 include ':instrumentation:lettuce:lettuce-5.1'
